fix: resolve 3 P1 from Codex v1.0 gate review
P1-1: Department identity collision fixed
  - DB lookup now includes course_name + duration_years
  - duration_years changed from Integer to Numeric(3,1) for fractional values (1.5, 2.4)
  - 学科別 diff improved from -301 to -264

P1-2: Step 9 parser rewritten by agent (complete overhaul)
  - School name: multi-pattern extraction (学校名, 大学等の名称)
  - Fiscal year: handles full-width date format
  - Department names: token-based course/dept separation
  - Graduation data: handles separate header+data rows

P1-3: PDF->DB pipeline wired
  - src/eidp/pipeline/ingest.py: parse_pdf -> department/department_yearly
  - CLI: `eidp ingest-pdfs` processes Document rows
  - Tested: 9 PDFs processed, 7 departments + 9 yearly rows created

P2: Optional dependency guard on discover-urls, ingest-pdfs commands

Codex v1.0 gate: 3x[P1] + 2x[P2] fixed.

+++ src/eidp/db/models.py
@@ -107,7 +107,7 @@ class Department(Base):
     course_name: Mapped[str | None] = mapped_column(String(200))  # 課程名
     canonical_name: Mapped[str] = mapped_column(String(200), nullable=False)
     course_type: Mapped[str | None] = mapped_column(String(10))
-    duration_years: Mapped[int | None] = mapped_column(Integer)
+    duration_years: Mapped[float | None] = mapped_column(Numeric(3, 1))
     field_category: Mapped[str | None] = mapped_column(String(50))
     status: Mapped[str] = mapped_column(String(20), default="active")
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
